:heavy_plus_sign: Add FieldNames class
This also marks the beginning of removing support for Python 2 (closes #128)
